+++ packages/Peptides/src/utils/misc.ts
@@ -1,11 +1,8 @@
-import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 import * as C from './constants';
 import * as type from './types';
 import {StringDictionary} from '@datagrok-libraries/utils/src/type-declarations';
-import {getSplitter} from '@datagrok-libraries/bio/src/utils/macromolecule/utils';
-import {TAGS as bioTags} from '@datagrok-libraries/bio/src/utils/macromolecule';
 
 export function getTypedArrayConstructor(
   maxNum: number): Uint8ArrayConstructor | Uint16ArrayConstructor | Uint32ArrayConstructor {